all library crates compile on thumbv7em-none-eabihf with embassy
- Update tinyvec to no_std-compatible version (default-features = false,
  features = ["alloc"])
- Add getrandom/custom to embassy feature (integrator provides RNG)
- Set default-features = false on workspace deps: thiserror, bitvec,
  hex, ux, num-traits, and all internal zwave-* crates
- CLI explicitly enables std features on all dependencies
- Add Debug derive to embassy Instant wrapper
- Replace f32::round() with integer arithmetic (no_std compatible)
- Use AtomicI32 instead of AtomicI64 (not available on 32-bit ARM)

Verified: all 5 lib crates compile on thumbv7em-none-eabihf with
--no-default-features --features embassy. Full std build and all
53 tests pass.

+++ packages/driver/src/driver/awaited.rs
@@ -1,6 +1,6 @@
 use crate::error::{Error, Result};
 use zwave_pal::prelude::*;
-use core::sync::atomic::{AtomicI64, Ordering};
+use core::sync::atomic::{AtomicI32, Ordering};
 use zwave_pal::time::MaybeSleep;
 use zwave_pal::channel::oneshot;
 
@@ -12,14 +12,14 @@ pub type Predicate<T> = Box<dyn Fn(&T) -> bool + Sync + Send>;
 /// Adding an entry hands out an `AwaitedRef`, which is used to receive the value when it is
 /// available. The `AwaitedRef` is automatically removed from the registry when it is dropped.
 pub struct AwaitedRegistry<T> {
-    next_id: AtomicI64,
+    next_id: AtomicI32,
     store: zwave_pal::sync::Mutex<Vec<Awaited<T>>>,
 }
 
 impl<T> Default for AwaitedRegistry<T> {
     fn default() -> Self {
         Self {
-            next_id: AtomicI64::new(0),
+            next_id: AtomicI32::new(0),
             store: zwave_pal::sync::Mutex::default(),
         }
     }
@@ -63,13 +63,13 @@ impl<T> AwaitedRegistry<T> {
 }
 
 pub struct Awaited<T> {
-    pub id: i64,
+    pub id: i32,
     pub predicate: Predicate<T>,
     pub channel: oneshot::Sender<T>,
 }
 
 pub struct AwaitedRef<T> {
-    id: i64,
+    id: i32,
     registry: Arc<AwaitedRegistry<T>>,
     timeout: Option<core::time::Duration>,
     channel: Option<oneshot::Receiver<T>>,
@@ -77,7 +77,7 @@ pub struct AwaitedRef<T> {
 
 impl<T> AwaitedRef<T> {
     pub fn new(
-        id: i64,
+        id: i32,
         registry: Arc<AwaitedRegistry<T>>,
         timeout: Option<core::time::Duration>,
         channel: oneshot::Receiver<T>,
